requirements/base.txt
chardet==5.2.0
cryptography==39.0.1
deep-translator==1.8.3
deprecated==1.2.13
django==5.0.1
django-admin-inline-paginator==0.4.0
django-admin-interface==0.28.3
django-appconf==1.0.5
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-health-check==3.19.0
django-import-export==2.8.0
django-ipware==5.0.0
django-ses==4.4.0
django-storages==1.13.1
django-tinymce==3.4.0
djangorestframework==3.15.1
drf-spectacular==0.28.0
gevent==25.5.1
gunicorn==23.0.0
hiredis==2.2.3
httplib2==0.20.4
markuppy==1.14
mysql-connector-python==9.2.0
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.2
psutil==7.0.0
pymongo==4.11.1
psycopg2==2.9.3
python-http-client==3.3.7
redis==5.0.0
requests==2.31.0
requests-toolbelt==0.9.1
tenacity==9.0.0
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0

# Authentication dependencies
python-jwt==2.0.1
djangorestframework-simplejwt==5.3.1

# AWS SDK for Python
boto3==1.35.18
urllib3<2.0
